Add chucker to code
