--- common/src/main/java/com/can_apps/common/CommonHttpClient.kt
@@ -1,35 +1,0 @@
-package com.can_apps.common
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-interface CommonHttpClient {
-
-    fun buildRank(): Retrofit
-
-    fun buildBad(): Retrofit
-}
-
-internal class Urls {
-    companion object {
-        const val RANK_URL = "https://www.stairwaylearning.com"
-        const val BAD_URL = "https://breakingbadapi.com"
-    }
-}
-
-class CommonHttpClientProvider : CommonHttpClient {
-
-    override fun buildRank(): Retrofit =
-        Retrofit
-            .Builder()
-            .baseUrl(Urls.RANK_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-    override fun buildBad(): Retrofit =
-        Retrofit
-            .Builder()
-            .baseUrl(Urls.BAD_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-}
