+++ common/src/main/java/com/can_apps/common/wrappers/CommonCalendarWrapper.kt
@@ -1,4 +1,4 @@
-package com.can_apps.common
+package com.can_apps.common.wrappers
 
 import java.util.Calendar
 import java.util.Locale
