+++ features/rank_board/src/main/java/com/can_apps/rank_board/core/RankInteractor.kt
@@ -1,7 +1,7 @@
 package com.can_apps.rank_board.core
 
-import com.can_apps.common.CommonCalendarWrapper
-import com.can_apps.common.CommonCoroutineDispatcherFactory
+import com.can_apps.common.wrappers.CommonCalendarWrapper
+import com.can_apps.common.coroutines.CommonCoroutineDispatcherFactory
 import kotlin.coroutines.CoroutineContext
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
