+++ features/rank_board/src/test/java/com/can_apps/rank_board/RankIntegrationTest.kt
@@ -1,7 +1,7 @@
 package com.can_apps.rank_board
 
-import com.can_apps.common.CommonCalendarWrapper
-import com.can_apps.common.CommonStringResource
+import com.can_apps.common.wrappers.CommonCalendarWrapper
+import com.can_apps.common.wrappers.CommonStringResource
 import com.can_apps.rank_board.bresenter.RankModel
 import com.can_apps.rank_board.bresenter.RankUsernameModel
 import com.can_apps.rank_board.bresenter.RankXpModel
