+++ features/rank_board/src/test/java/com/can_apps/rank_board/bresenter/RankPresenterTest.kt
@@ -1,7 +1,7 @@
 package com.can_apps.rank_board.bresenter
 
-import com.can_apps.common.CommonCoroutineDispatcherFactory
-import com.can_apps.common.CommonCoroutineDispatcherFactoryUnconfined
+import com.can_apps.common.coroutines.CommonCoroutineDispatcherFactory
+import com.can_apps.common.coroutines.CommonCoroutineDispatcherFactoryUnconfined
 import com.can_apps.rank_board.core.RankContract
 import com.can_apps.rank_board.core.RankDomain
 import io.mockk.MockKAnnotations
